requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
blinker==1.9.0
cachetools==5.5.2
certifi==2026.2.25
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
greenlet==3.3.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
Jinja2==3.1.6
jiter==0.13.0
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
langchain==0.3.26
langchain-community==0.3.26
langchain-core==0.3.83
langchain-openai==0.3.27
langchain-text-splitters==0.3.11
langsmith==0.7.16
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
multidict==6.7.1
mypy_extensions==1.1.0
narwhals==2.18.0
numpy==2.4.3
openai==1.91.0
orjson==3.11.7
packaging==24.2
pandas==2.3.3
pillow==11.3.0
propcache==0.4.1
protobuf==5.29.6
pyarrow==23.0.1
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
pydeck==0.9.1
Pygments==2.19.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
pytz==2026.1.post1
PyYAML==6.0.3
referencing==0.37.0
regex==2026.2.28
requests==2.32.5
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.30.0
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
SQLAlchemy==2.0.48
streamlit==1.41.1
tenacity==9.1.4
tiktoken==0.12.0
toml==0.10.2
tornado==6.5.5
tqdm==4.67.3
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uuid_utils==0.14.1
watchdog==6.0.0
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
